+++ tasks/seq/savchenko_m_min_matrix/perf_tests/main_savchenko.cpp
@@ -6,7 +6,7 @@
 #include <vector>
 
 #include "core/perf/include/perf.hpp"
-#include "seq/savchenko_m_min_matrix/include/ops_seq.hpp"
+#include "seq/savchenko_m_min_matrix/include/ops_seq_savchenko.hpp"
 
 TEST(savchenko_m_min_matrix_seq, test_pipeline_run) {
   std::vector<int> matrix;
@@ -29,7 +29,8 @@ TEST(savchenko_m_min_matrix_seq, test_pipeline_run) {
 
   // Create TaskData
   taskDataSeq->inputs.emplace_back(reinterpret_cast<uint8_t *>(matrix.data()));
-  taskDataSeq->inputs_count.emplace_back(matrix.size());
+  taskDataSeq->inputs_count.emplace_back(rows);
+  taskDataSeq->inputs_count.emplace_back(columns);
 
   taskDataSeq->outputs.emplace_back(reinterpret_cast<uint8_t *>(min_value.data()));
   taskDataSeq->outputs_count.emplace_back(min_value.size());
@@ -78,7 +79,8 @@ TEST(savchenko_m_min_matrix_seq, test_task_run) {
 
   // Create TaskData
   taskDataSeq->inputs.emplace_back(reinterpret_cast<uint8_t *>(matrix.data()));
-  taskDataSeq->inputs_count.emplace_back(matrix.size());
+  taskDataSeq->inputs_count.emplace_back(rows);
+  taskDataSeq->inputs_count.emplace_back(columns);
 
   taskDataSeq->outputs.emplace_back(reinterpret_cast<uint8_t *>(min_value.data()));
   taskDataSeq->outputs_count.emplace_back(min_value.size());
